image_storage/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
minio==7.2.7
python-dotenv==1.0.1
